requirements.txt
streamlit
opencv-python
numpy
pillow
requests
mahotas
face_recognition
deepface
scikit-image   # Pour GLCM
